refactor. ChatMember Controller, Service -> ChatRoom Controller, Service
