--- src/main/java/com/example/gotogetherbe/chat/controller/ChatMemberController.java
@@ -1,28 +1,0 @@
-package com.example.gotogetherbe.chat.controller;
-
-import com.example.gotogetherbe.auth.config.LoginUser;
-import com.example.gotogetherbe.chat.dto.ChatMemberDto;
-import com.example.gotogetherbe.chat.service.ChatMemberService;
-import com.example.gotogetherbe.global.util.jwt.JwtUserDetails;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class ChatMemberController {
-  private final ChatMemberService chatMemberService;
-
-  // 채팅방 참여자 목록 조회
-  @GetMapping("/api/chat-member/list/{chatRoomId}")
-  public ResponseEntity<List<ChatMemberDto>> getChatMemberList(
-      @LoginUser String username,
-      @PathVariable Long chatRoomId
-  ) {
-    return ResponseEntity.ok(chatMemberService.getChatMemberList(username,chatRoomId));
-  }
-}
