--- src/test/api-test/ChatMemberTest.http
@@ -1,4 +1,0 @@
-### 채팅방 참여자 목록 조회
-POST http://localhost:8080/api/chat-member/list/{chatRoomId}
-Content-Type: application/json
-Authorization: Bearer
