+++ src/test/api-test/ChatRoomTest.http
@@ -1,17 +1,23 @@
 ### 채팅방 생성
 POST http://localhost:8080/api/chat-room/1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3QxMjM0QGdtYWlsLmNvbSIsImlhdCI6MTcxMDA5ODI3OCwiZXhwIjoxNzEwMTAwMDc4fQ.WdULY0syhLbyG0VgFhQ272_342RYZ-eALrK7E8rstXI
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3QxMjM0QGdtYWlsLmNvbSIsImlhdCI6MTcxMDI1MTA5OCwiZXhwIjoxNzEwMjUyODk4fQ.CfrHnLZO2lIlH3sIlk_utNwKfymSc4VwhVVNYo9PfAo
 
 ### 참여중인 채팅방 목록 조회
-GET http://localhost:8080/api/chat-room/list
+GET http://localhost:8080/api/chat-room/my-list
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3QxMjM0QGdtYWlsLmNvbSIsImlhdCI6MTcxMDA5ODI3OCwiZXhwIjoxNzEwMTAwMDc4fQ.WdULY0syhLbyG0VgFhQ272_342RYZ-eALrK7E8rstXI
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3Q1Njc4QGdtYWlsLmNvbSIsImlhdCI6MTcxMDI1MTEwNiwiZXhwIjoxNzEwMjUyOTA2fQ.imTqtncgnO7QD9oD8-cuDIk95nPDuZ9LGlvFqLsXtZY
+
+### 채팅방에 참여중인 멤버 목록 조회
+GET http://localhost:8080/api/chat-room/member-list/1
+Content-Type: application/json
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3Q1Njc4QGdtYWlsLmNvbSIsImlhdCI6MTcxMDI1MTEwNiwiZXhwIjoxNzEwMjUyOTA2fQ.imTqtncgnO7QD9oD8-cuDIk95nPDuZ9LGlvFqLsXtZY
+
 
 ### 채팅방 입장
 POST http://localhost:8080/api/chat-room/enter/1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3QxMjM0QGdtYWlsLmNvbSIsImlhdCI6MTcxMDA5ODI3OCwiZXhwIjoxNzEwMTAwMDc4fQ.WdULY0syhLbyG0VgFhQ272_342RYZ-eALrK7E8rstXI
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3Q1Njc4QGdtYWlsLmNvbSIsImlhdCI6MTcxMDI1MTEwNiwiZXhwIjoxNzEwMjUyOTA2fQ.imTqtncgnO7QD9oD8-cuDIk95nPDuZ9LGlvFqLsXtZY
 
 ### 채팅방 퇴장
 DELETE http://localhost:8080/api/chat-room/exit/1
@@ -21,4 +27,4 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlc
 ### 채팅방 메세지 조회
 GET http://localhost:8080/api/chat-room/message/1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3QxMjM0QGdtYWlsLmNvbSIsImlhdCI6MTcxMDA5ODI3OCwiZXhwIjoxNzEwMTAwMDc4fQ.WdULY0syhLbyG0VgFhQ272_342RYZ-eALrK7E8rstXI
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhY2Nlc3MtdG9rZW4iLCJyb2xlcyI6IlJPTEVfVVNFUiIsInVzZXJJZCI6InRlc3Q1Njc4QGdtYWlsLmNvbSIsImlhdCI6MTcxMDI1MTEwNiwiZXhwIjoxNzEwMjUyOTA2fQ.imTqtncgnO7QD9oD8-cuDIk95nPDuZ9LGlvFqLsXtZY
